Moderate changes and additions
The code now calculates the polarisation centrally in calc_polarization.f90 and provides the Ek, Bk, and Avec polarisation vectors from that subroutine. This reduces redundancy. 
The calculation of the wave energy has been corrected.
The code makes additional outputs about the contributions of all species to gamma, and of the contribution of all species to the heating rate (this goes to the _wave output file).