Major code update
The code can now handle an arbitrary number of species (not limited to 10 anymore).
It also has the ability to calculate susceptibilities based on cold-plasma theory. This feature is activated if beta is set to 0.d0 for a given species.
The previous input files do not work anymore since the namelist structure needed to be changed (updated example .in files included in folder).